29-09-2021 Perform delete, update operation using both put & patch and also using map for particular change; Add Base model & common dto

+++ src/main/java/com/example/todos/taskapp/Repository/TaskRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TaskRepository extends JpaRepository<TaskModel, Integer> {
+public interface TaskRepository extends JpaRepository<TaskModel, Long> {
 }
